pyproject.toml
[tool.poetry]
name = "crowdeval"
version = "0.1.0"
description = ""
authors = ["Freddy Heppell <freddy@freddyheppell.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
environs = "^9.3.1"
Flask-Static-Digest = "^0.2.1"
Flask-Dance = {extras = ["sqla"], version = "^3.2.0"}
Flask-SQLAlchemy = "^2.4.4"
Flask-Login = "^0.5.0"
Flask-Migrate = "^2.6.0"
blinker = "^1.4"
Flask-DebugToolbar = "^0.11.0"
mysql-connector-python = "^8.0.23"
Flask-WTF = "^0.14.3"
Flask-Bootstrap = "^3.3.7"
TwitterAPI = "^2.6.9"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pylint = "^2.6.0"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-blind-except = "^0.2.0"
flake8-debugger = "^4.0.0"
flake8-docstrings = "^1.5.0"
flake8-isort = "^4.0.0"
isort = "^5.7.0"
pep8-naming = "^0.11.1"
WebTest = "^2.0.35"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip_glob = ["migrations/*"]